refactor: extract subagent rehydrate logic
